Delete BrushComponent.tsx
