+++ src/stories/cartesian/brush/Brush.stories.tsx
@@ -1,15 +1,81 @@
 import React from 'react';
-import BrushComponent from './BrushComponent';
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { ComponentMeta, ComponentStory } from '@storybook/react';
 import { dateData } from '../../data';
+import { Surface, Brush } from '../../..';
 
 export default {
   title: 'Recharts/Brush',
-  component: BrushComponent,
-};
+  component: Brush,
+} as ComponentMeta<typeof Brush>;
 
-function Template(args: any) {
-  return <BrushComponent {...args} />;
-}
+const Template: ComponentStory<typeof Brush> = args => {
+  const [simple, setSimple] = React.useState({
+    startIndex: 0,
+    endIndex: args.data.length - 1,
+  });
+
+  const [gap, setGap] = React.useState({
+    startIndex: 0,
+    endIndex: args.data.length - 1,
+  });
+
+  const handleChange = (res: any) => {
+    setSimple(res);
+  };
+
+  const handleGapChange = (res: any) => {
+    setGap(res);
+  };
+
+  // eslint-disable-next-line no-shadow
+  const renderTraveller = (props: { x: number; y: number; width: number; height: number }) => {
+    const { x, y, width, height } = props;
+
+    return (
+      <path
+        d={`M${x + width / 2},${y}L${x + width},${y + height / 2}L${x + width / 2},${y + height}L${x},${
+          y + height / 2
+        }Z`}
+        fill="red"
+        stroke="none"
+      />
+    );
+  };
+
+  return (
+    <div>
+      <p>Simple Brush</p>
+      <Surface width={800} height={200}>
+        <Brush
+          startIndex={simple.startIndex}
+          endIndex={simple.endIndex}
+          x={100}
+          y={50}
+          width={400}
+          height={40}
+          data={args.data}
+          onChange={handleChange}
+          traveller={renderTraveller}
+        />
+      </Surface>
+      <p>Brush has specified gap</p>
+      <Surface width={800} height={200}>
+        <Brush
+          startIndex={gap.startIndex}
+          endIndex={gap.endIndex}
+          x={100}
+          y={50}
+          width={400}
+          height={40}
+          data={args.data}
+          gap={5}
+          onChange={handleGapChange}
+        />
+      </Surface>
+    </div>
+  );
+};
 
 export const Simple = Template.bind({});
 Simple.args = {
